docs/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Document
from .serializers import DocumentSerializer

# View para a página de documentos
def documents_list_view(request):
    return render(request, 'documents.html')

class DocumentViewSet(viewsets.ModelViewSet):
    queryset = Document.objects.all()
    serializer_class = DocumentSerializer